+++ app/lib/bk/compat/parsers/circleci/jobs.rb
@@ -38,7 +38,7 @@ module BK
       end
 
       def parse_command(name, config)
-        BK::Compat::CircleCISteps::CircleCIStep.new(key: name).tap do |bk_step|
+        BK::Compat::CircleCIStep.new(key: name).tap do |bk_step|
           translate_steps(config['steps']).each { |step| bk_step << step }
           bk_step.parameters = config.fetch('parameters', {})
         end
