+++ app/lib/bk/compat/parsers/gha.rb
@@ -3,6 +3,8 @@
 require_relative '../error'
 require_relative '../translator'
 require_relative '../models/pipeline'
+require_relative '../models/steps/gha/gha_step'
+
 require_relative 'gha/branches'
 require_relative 'gha/jobs'
 require_relative 'gha/steps'
