+++ app/lib/bk/compat/parsers/gha/steps.rb
@@ -43,31 +43,5 @@ module BK
         skip ? arr : ['('] + arr + [')']
       end
     end
-
-    # wrapper class to setup parameters
-    class GHAStep < BK::Compat::CommandStep
-      attr_accessor :transformer
-
-      def initialize(*, **)
-        super
-        @transformer = method(:gha_params)
-      end
-
-      EXPRESSION_REGEXP = /\${{\s*(?<data>.*)\s*}}/
-      EXPRESSION_TRANSFORMER = BK::Compat::GithubExpressions.new
-
-      def gha_params(value)
-        value.gsub(EXPRESSION_REGEXP) do |v|
-          d = EXPRESSION_REGEXP.match(v)
-          EXPRESSION_TRANSFORMER.transform(d[:data])
-        end
-      end
-
-      def to_h
-        @transformer = nil
-
-        super
-      end
-    end
   end
 end
